Preserve field party index width
Keep the selected party index as a signed machine-width argument through field model creation, matching the retail ABI lifetime and avoiding the candidate's early zero-extension path.

Constraint: Genuine C only; do not reproduce the rejected stack-local resource table copy.
Confidence: high
Scope-risk: narrow
Directive: The s32 parameter is intentional even though downstream party access accepts the same numeric range.
Tested: Exact active wrapper build; focused fn_801E075C 45.049690% to 48.496895%; git diff --check.
Not-tested: Runtime field model creation.

+++ src/game/field_range_801DF790.c
@@ -623,7 +623,7 @@ static const u32 sBallModels[13] = {
  * fn_801E075C - Create the field model used for the selected party Pokemon.
  * Address: 0x801E075C | Size: 0x284
  */
-void fn_801E075C(u16 partyIndex)
+void fn_801E075C(s32 partyIndex)
 {
     extern u8 pokemonBiosGetCatchBallId(void* pokemon);
     extern u16 pokemonBiosGetPokemonDataId(void* pokemon);
